+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/Components.kt
@@ -158,7 +158,7 @@ class Components(private val context: Context) {
         AddonManager(core.store, core.engine, addonsProvider, addonUpdater)
     }
 
-    val analytics by lazyMonitored { Analytics(context) }
+    val analytics by lazyMonitored { Analytics(context, performance.visualCompletenessQueue.queue) }
     val publicSuffixList by lazyMonitored { PublicSuffixList(context) }
     val clipboardHandler by lazyMonitored { ClipboardHandler(context) }
     val performance by lazyMonitored { PerformanceComponent() }